MPKG_Embedding/requirements.txt
pandas~=2.0.3
py2neo~=2021.2.4
numpy~=1.24.4
tqdm~=4.66.5
requests~=2.32.3
scikit-learn~=1.3.2
pyyaml~=6.0.2
click~=8.1.7